more error fixing
yep i genuinely cant code
had an overflow error, also changed up the datatype of the position to allow for bigger numbers

+++ src/vertex/mod.rs
@@ -1,5 +1,5 @@
 impl Vertex {
-    pub fn new(pos: [i8; 3], tc: [i8; 2]) -> Vertex {
+    pub fn new(pos: [i16; 3], tc: [i8; 2]) -> Vertex {
         Vertex {
             a_pos: [pos[0], pos[1], pos[2], 1],
             a_tex_coord: tc,
@@ -8,6 +8,6 @@ impl Vertex {
 }
 
 gfx_vertex_struct!( Vertex {
-    a_pos: [i8; 4] = "a_pos",
+    a_pos: [i16; 4] = "a_pos",
     a_tex_coord: [i8; 2] = "a_tex_coord",
 });
